Move all of the client code into the src/ directory to match the structure of services/
