+++ client/src/lib/__mocks__/fs.js
