+++ client/src/lib/ping.js
