+++ client/src/lib/rand-patch.js
