+++ client/src/lib/registration.js
