+++ client/test/client.test.js
@@ -1,5 +1,5 @@
 const assert = require('assert');
-const client = require('../client');
+const client = require('../src/client');
 
 describe('The base client module', () => {
   it('should interpret properly', () => {
