+++ client/test/registration.test.js
@@ -3,7 +3,7 @@ jest.mock('fs');
 const assert       = require('assert');
 const fs           = require('fs');
 const path         = require('path');
-const Registration = require('../lib/registration');
+const Registration = require('../src/lib/registration');
 
 describe('The registration module', () => {
   beforeEach(() => {
